pyproject.toml
[project]
name = "auth-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "augdev1",email = "augustosousa1798@gmail.com"}
]
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.134.0,<0.135.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "bcrypt (==3.2.2)",
    "pyotp (>=2.9.0,<3.0.0)",
    "qrcode[pil] (>=8.2,<9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
